remove condition for bhj to avoid timeout

Created-by: liuyu 00809192
Author-id: 53630
MR-id: 21975490
Commit-by: liuyu
Merged-by: ruanrunxue 00424934
E2E-issues: issue653
Description: + [ ] I've added/changed the documentation
+ [ ] I've added test for the code

See merge request: Kunpeng-Computing-DC/BigData/OmniOperatorJIT!808

+++ core/test/vector/omni_row_test.cpp
@@ -217,7 +217,7 @@ TEST(omni_row, fill_buffer_and_deserial_to_vector)
 
     VectorBatch *result = CreateVectorBatch(dataTypes, rowNumber, fakeData1, fakeData2, fakeData3, fakeData4);
     BaseVector *vecs[types.size()];
-    for (int32_t i = 0; i < types.size(); ++i) {
+    for (int32_t i = 0; i < static_cast<int32_t>(types.size()); ++i) {
         vecs[i] = result->Get(i);
     }
 
@@ -317,7 +317,7 @@ TEST(omni_row, fill_bool_buffer_and_deserial_to_vector)
 
     VectorBatch *result = CreateVectorBatch(dataTypes, rowNumber, fakedata1);
     BaseVector *vecs[types.size()];
-    for (int32_t i = 0; i < types.size(); ++i) {
+    for (int32_t i = 0; i < static_cast<int32_t>(types.size()); ++i) {
         vecs[i] = result->Get(i);
     }
